% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beavers}
\alias{beavers}
\title{Beavers}
\format{
A dataframe with ... rows and ... columns
\describe{
\item{column1}{what's in the first column}
\item{column2}{what's in the second column}
}
}
\source{
A url for where the data came from.
}
\usage{
beavers
}
\description{
Data on Beavers that Glen found. TODO: Glen to write this documentation. Also,
Glen to include only rows and columns that are actually needed in the dataset.
}
\keyword{datasets}
