% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{flatten_bdmat}
\alias{flatten_bdmat}
\title{Flatten a block-diagonal sparse matrix}
\usage{
flatten_bdmat(Z, n, d)
}
\arguments{
\item{Z}{Block diagonal matrix with equal-sized blocks (don't have to be square)}

\item{m}{Number of rows in each block}

\item{p}{Number of columns in each block}
}
\value{
A dense matrix with the appropriate structure
}
\description{
Take a n x d block diagonal matrix with blocks of dimension m x p
and flatten to a (dense) matrix of dimension n x p
}
