% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{logit}
\alias{logit}
\alias{ilogit}
\title{Logit and ilogit}
\usage{
logit(x)

ilogit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
\code{logit}: the logit of \code{x}, \code{log(x/(1-x))}

\code{ilogit}: the inverse logit of \code{x}, \code{1/(1+exp(-x))}
}
\description{
Quick implementations of the logit and inverse logit functions
}
\examples{
logit(seq(0,1,by=.1))
ilogit(seq(-5,5,by=1))
}
