% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{newsparsemat}
\alias{newsparsemat}
\title{Create an emprty sparse matrix}
\usage{
newsparsemat(n, m)
}
\arguments{
\item{n}{Number of rows}

\item{m}{Number of columns}
}
\value{
An object of class \code{dgCMatrix} having the appropriate structure,
and no data in teh \code{x} slot.
}
\description{
Wrapper for new('dgCMatrix',...). Creates an empty sparse \code{dgCMatrix}
of specified dimension.
}
\examples{
newsparsemat(10,5)

}
