% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mam.R
\name{mam}
\alias{mam}
\alias{mam.default}
\alias{mam.gam}
\title{Estimate marginal coefficients from an additive model}
\usage{
mam(mod, ...)

\method{mam}{default}(mod, ...)

\method{mam}{gam}(mod, ...)
}
\arguments{
\item{mod}{Object of class \code{gam} created using \code{mgcv::gam}.}

\item{...}{not used.}
}
\value{
Object of classes \code{c('mam','gam','glm','lm')} representing the marginal additive model.
}
\description{
This function takes an object of class \code{gam} created using \code{mgcv::gam},
and returns a marginal additive model object.
}
\examples{
## Not run

}
